fix: Correct multiple warnings in MinGW64/MinGW32
In MinGW64/MinGW32 environment, gcc will give warnings when we use 'char'
as an array subscript, and when we use C11 functions, such as getenv_s().

We use explicit casting to resolve the array subscript issues, and check
if the build environment supports C11 before calling C11 functions.

resolves #61

+++ src/arg_date.c
@@ -262,7 +262,7 @@ char* arg_strptime(const char* buf, const char* fmt, struct tm* tm) {
 
         /* Eat up white-space. */
         if (isspace(c)) {
-            while (isspace(*bp))
+            while (isspace((int)(*bp)))
                 bp++;
 
             fmt++;
@@ -534,7 +534,7 @@ char* arg_strptime(const char* buf, const char* fmt, struct tm* tm) {
             case 'n': /* Any kind of white-space. */
             case 't':
                 LEGAL_ALT(0);
-                while (isspace(*bp))
+                while (isspace((int)(*bp)))
                     bp++;
                 break;
